Rework: move duplicating functions from adapt registry to BaseOptimizationAdapter (2)
